Player and Solid Object Interaction
Also controls are guaranteed to work now! I’m pleased. :D

+++ Recursion?/Instances/Room 1/Room1Hazard.h
@@ -12,6 +12,6 @@
 /**@brief A hazard object that harms the player.*/
 @interface Room1Hazard : Instance
 /**@brief The frame of animation.*/
-@property int frame;
+@property float frame;
 @end
 #endif
